add nodes folder and move BitcoinPrice node to it
